pyproject.toml
[project]
name = "myconfbot"
version = "0.1.0"
description = "Telegram bot for confectioner"
readme = "README.md"
requires-python = ">=3.8.1"
license = {text = "MIT"}
authors = [
    {name = "Serg Tumali", email = "setum77@gmail.com"}
]
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "Pillow>=9.0.0",
    "psycopg2-binary>=2.9.0",
    "pytelegrambotapi>=4.0.0",
    "requests>=2.28.0",
    "SQLAlchemy>=2.0.0",
]

[project.scripts]
confectionery-bot = "src.myconfbot.__main__:main"

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "isort>=5.12",
    "pytest>=7.0",
    "pytest-asyncio>=0.20",
    "python-multipart>=0.0.5",
    "flake8>=6.0",
    "mypy>=1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"